requirements.txt
absl-py==1.0.0
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.0.3
antlr4-python3-runtime==4.8
appnope==0.1.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
black==22.1.0
blis==0.7.6
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
dataclasses==0.6
decorator==5.1.1
docker-pycreds==0.4.0
docstring-parser==0.13
effdet==0.2.4
efficientnet-pytorch==0.6.3
entrypoints==0.4
fastai==2.5.3
fastapi==0.68.2
fastcore==1.3.29
fastdownload==0.0.5
fastprogress==1.0.2
fire==0.4.0
fonttools==4.31.1
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
fvcore==0.1.5.post20220305
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h11==0.13.0
icevision==0.12.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
iopath==0.1.9
ipykernel==5.5.6
ipython==7.32.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonargparse==4.4.0
jsonlines==3.0.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.0
kornia==0.6.4
langcodes==3.3.0
lightning-bolts==0.5.0
lightning-flash==0.7.1
loguru==0.6.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
multidict==6.0.2
munch==2.5.0
murmurhash==1.0.6
mypy-extensions==0.4.3
nest-asyncio==1.5.4
networkx==2.6.3
nose==1.3.7
numpy==1.21.5
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pathy==0.6.1
peewee==3.14.10
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plac==1.1.3
platformdirs==2.5.1
portalocker==2.4.0
preshed==3.0.6
pretrainedmodels==0.7.4
prodigy==1.11.7
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydantic==1.8.2
pyDeprecate==0.3.1
Pygments==2.11.2
PyJWT==2.3.0
pyngrok==5.1.0
pyparsing==3.0.7
pystiche==1.0.1
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
resnest==0.0.6b20220321
rsa==4.8
sahi==0.9.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
segmentation-models-pytorch==0.2.1
sentry-sdk==1.5.8
setproctitle==1.2.2
Shapely==1.8.1.post1
shortuuid==1.0.8
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
starlette==0.14.2
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminaltables==3.1.10
thinc==8.0.15
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
tomli==2.0.1
toolz==0.11.2
torch==1.10.2
torchmetrics==0.7.2
torchvision==0.11.3
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typed-ast==1.5.2
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.9
uvicorn==0.13.4
wandb==0.12.11
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.3
yacs==0.1.8
yarl==1.7.2
yaspin==2.1.0
yolov5-icevision==6.0.0
zipp==3.7.0
